Update embind tests to run with the Closure Compiler
The tests themselves must not be run through the Closure Compiler, as it will
mangle references to the names exported through embind. Since embind itself
uses string literals to export its names, this would break. Therefore, the tests
and adapter are now manually concatenated to the output JavaScript post-Closure.

The IMVU test adapter also needed to be updated so that it does not rely on
the assert function already having been created by Emscripten. It appears that
this was its natural state at some point in history.

Note that the embind tests prove that embind is broken when run through the
Closure Compiler; embind will be fixed in the next commit :)